refactor(effis): use u64 not u128 for ids

+++ cli/src/main.rs
@@ -74,7 +74,7 @@ enum AttachmentSubcommand {
     /// Removes an attachment
     Remove {
         /// The id of the attchment to be removed
-        id: u128,
+        id: u64,
     },
 }
